single admin playlist & themes handler to context
