+++ src/components/admin/common/NumberOption.tsx
@@ -134,5 +134,5 @@ export function useNumberOption(settingKey: keyof SettingsList) {
     setValueState(value)
   }
 
-  return { value, setValue } as const
+  return { value, setValue }
 }
